safari-600.7-branch/LayoutTests/fast/dom/DOMURL/set-href-attribute-hostname-expected.txt

Test setting the hostname attribute of the URL in a URL object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Basic test
PASS a.href is 'https://www.otherdomain.com:8080/path/'
Extra slashes before hostname
PASS a.href is 'https://www.otherdomain.com:8080/path/'
Set hostname to URL with foo: protocol
PASS a.href is 'foo://www.otherdomain.com/path/'
Set hostname to null
PASS a.href is 'https://www.mydomain.com:8080/path/'
Set hostname to empty string
PASS a.href is 'https://www.mydomain.com:8080/path/'
Set hostname to URL with 2 colons
PASS a.href is 'really:bad:url'
Set a hostname that contains space in it
PASS a.href = "http://www.my domain.com/path/" threw exception TypeError: Type error.
PASS a.href is 'http://www.my domain.com/path/'
Set hostname on a local file
FAIL a.href should be c:/path/testurl.html. Was c://a/path/testurl.html.
Set hostname to undefined
PASS a.href is 'https://undefined:8080/path/'
PASS successfullyParsed is true

TEST COMPLETE
